# besoins/openapi.yaml
openapi: 3.1.0
info:
  title: Contribo - API de gestion associative
  version: 1.0.0-mvp
  description: |
    Contrat API du MVP Contribo. Les montants sont exprimés en francs guinéens (GNF),
    sous forme d'entiers sans sous-unité. L'API ne déclenche aucun paiement : elle
    enregistre uniquement des encaissements constatés par un utilisateur autorisé.
servers:
  - url: /api/v1
    description: Point d'entrée relatif du MVP
...
security:
  - bearerAuth: []
paths:
...
  /me:
    get:
      tags: [Espace personnel]
      summary: Consulter mon compte et mon profil membre
      operationId: getCurrentUser
      responses:
        '200':
          description: Compte et membre associés à la session.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentUser'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /me/dues:
    get:
      tags: [Espace personnel]
      summary: Consulter mes cotisations
      operationId: listMyDues
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/DueStatusFilter'
      responses:
        '200':
          description: Cotisations du membre connecté, de la plus récente à la plus ancienne.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DuePage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /me/contributions:
    get:
      tags: [Espace personnel]
      summary: Consulter mes contributions aux cagnottes
      operationId: listMyContributions
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Contributions rattachées au membre connecté, de la plus récente à la plus ancienne; les contributions externes sont exclues.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContributionPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /social-funds/{socialFundId}/contributions:
    get:
      tags: [Contributions]
      summary: Lister les contributions d'une cagnotte
      operationId: listSocialFundContributions
      parameters:
        - $ref: '#/components/parameters/SocialFundId'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/SearchQuery'
      responses:
        '200':
          description: Contributions de la cagnotte, de la plus récente à la plus ancienne.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContributionPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags: [Contributions]
      summary: Enregistrer une contribution à une cagnotte
      description: |
        Accessible aux Administrateurs, Trésoriers et Opérateurs dont l'autorisation globale est active.
        La contribution est un encaissement déjà constaté pour un membre ou un contributeur externe.
        La requête doit contenir exactement une identité contributrice. L'utilisateur et l'instant de
        saisie sont déterminés par le serveur. Une cagnotte clôturée refuse toute contribution.
      operationId: createContribution
      parameters:
        - $ref: '#/components/parameters/SocialFundId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateContributionRequest'
      responses:
        '201':
          description: Contribution enregistrée et bilan recalculé.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContributionCreationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /contributions:
    get:
      tags: [Contributions]
      summary: Lister les contributions aux cagnottes
      description: |
        Accessible aux Administrateurs, Trésoriers et Opérateurs. L'autorisation globale de
        l'Opérateur est requise pour écrire une contribution, pas pour consulter un historique
        autorisé. L'historique peut contenir des contributions rattachées à un membre ou à un
        contributeur externe.
      operationId: listContributions
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/SearchQuery'
        - $ref: '#/components/parameters/MemberIdQuery'
        - $ref: '#/components/parameters/SocialFundIdQuery'
      responses:
        '200':
          description: Contributions du plus récent au plus ancien.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContributionPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /contributions/{contributionId}:
    get:
      tags: [Contributions]
      summary: Consulter une contribution
      description: |
        Accessible aux rôles de gestion autorisés. Un Membre ne peut consulter que sa propre contribution.
      operationId: getContribution
      parameters:
        - $ref: '#/components/parameters/ContributionId'
      responses:
        '200':
          description: Contribution demandée.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contribution'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Jeton d'accès obtenu avec POST /auth/login.
  parameters:
    Page:
      name: page
      in: query
      description: Numéro de page indexé à partir de 0.
      schema:
        type: integer
        minimum: 0
        default: 0
      example: 0
    PageSize:
      name: size
      in: query
      description: Nombre maximal d'éléments retournés.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      example: 20
    SearchQuery:
      name: q
      in: query
      description: Recherche textuelle insensible à la casse sur les libellés visibles de la ressource.
      schema:
        type: string
        minLength: 1
        maxLength: 100
      example: Diallo
...
    DueStatusFilter:
      name: status
      in: query
      description: Filtre sur le statut de cotisation.
      schema:
        $ref: '#/components/schemas/DueStatus'
...
    SocialFundId:
      name: socialFundId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    ContributionId:
      name: contributionId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    MemberIdQuery:
      name: memberId
      in: query
      description: Limite les opérations à un membre.
      schema:
        type: string
        format: uuid
...
    SocialFundIdQuery:
      name: socialFundId
      in: query
      description: Limite les contributions à une cagnotte.
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Requête invalide ou validation échouée.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: VALIDATION_ERROR
            message: La requête contient des valeurs invalides.
            fieldErrors:
              - field: endDate
                code: DATE_RANGE_INVALID
                message: La date de fin doit être postérieure ou égale à la date de début.
    Unauthorized:
      description: Authentification absente ou invalide.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: AUTHENTICATION_REQUIRED
            message: Une authentification valide est requise.
    Forbidden:
      description: Le rôle ou l'autorisation du compte ne permet pas cette action.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: ACCESS_DENIED
            message: Vous n'êtes pas autorisé à effectuer cette action.
    NotFound:
      description: Ressource introuvable ou non visible par le compte connecté.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: RESOURCE_NOT_FOUND
            message: La ressource demandée est introuvable.
    Conflict:
      description: L'action viole une règle métier ou l'état courant de la ressource.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            overpayment:
              value:
                code: PAYMENT_EXCEEDS_REMAINING_AMOUNT
                message: Le montant dépasse le reste à payer.
            closedResource:
              value:
                code: RESOURCE_CLOSED
                message: La ressource est clôturée et n'accepte plus cette action.
...
    InternalServerError:
      description: Erreur interne non prévue.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: INTERNAL_ERROR
            message: Une erreur interne est survenue.
  schemas:
...
    UserRole:
      type: string
      description: Rôle applicatif, indépendant de la fonction associative.
      enum: [ADMINISTRATOR, TREASURER, OPERATOR, MEMBER]
      example: OPERATOR
    MemberStatus:
      type: string
      enum: [ACTIVE, INACTIVE]
      example: ACTIVE
    CampaignStatus:
      type: string
      description: UPCOMING tant que la campagne n'est pas ouverte explicitement, OPEN après son ouverture et jusqu'à la clôture explicite, CLOSED après clôture.
      enum: [UPCOMING, OPEN, CLOSED]
      example: OPEN
    DueStatus:
      type: string
      description: |
        DUE si aucun règlement n'existe, PARTIALLY_PAID si le total payé est intermédiaire,
        PAID si le reste vaut zéro, OVERDUE si l'échéance est dépassée avec un reste positif.
      enum: [DUE, PARTIALLY_PAID, PAID, OVERDUE]
      example: PARTIALLY_PAID
    SocialFundStatus:
      type: string
      enum: [OPEN, CLOSED]
      example: OPEN
    PaymentMethod:
      type: string
      enum: [CASH, MOBILE_MONEY, BANK_TRANSFER]
      example: MOBILE_MONEY
    SocialEventType:
      type: string
      enum: [WEDDING, BAPTISM, DEATH, BIRTH, OTHER]
      example: WEDDING
    CurrencyCode:
      type: string
      enum: [GNF]
      example: GNF
    CurrentUser:
      type: object
      additionalProperties: false
      required: [userId, association, member, role, operatorCanRecordPayments, accountActive]
      properties:
        userId:
          type: string
          format: uuid
          readOnly: true
        association:
          $ref: '#/components/schemas/AssociationSummary'
        member:
          $ref: '#/components/schemas/MemberSummary'
        role:
          $ref: '#/components/schemas/UserRole'
        operatorCanRecordPayments:
          type: boolean
          description: Attribut significatif uniquement lorsque role vaut OPERATOR; false sinon.
          example: true
        accountActive:
          type: boolean
          readOnly: true
          example: true
...
    PersonSummary:
      type: object
      additionalProperties: false
      required: [id, displayName]
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        displayName:
          type: string
          minLength: 1
          maxLength: 201
          readOnly: true
          example: Amadou Diallo
    AssociationSummary:
      type: object
      additionalProperties: false
      required: [id, name, currency]
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        name:
          type: string
          minLength: 1
          maxLength: 150
          readOnly: true
          example: Union Contribo Conakry
        currency:
          $ref: '#/components/schemas/CurrencyCode'
    AuditActor:
      type: object
      additionalProperties: false
      required: [userId, displayName]
      properties:
        userId:
          type: string
          format: uuid
          readOnly: true
        displayName:
          type: string
          minLength: 1
          maxLength: 201
          readOnly: true
          example: Aminata Touré
    IncomeCategorySummary:
      type: object
      required: [id, label]
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        label:
          type: string
          minLength: 1
          maxLength: 100
          example: Catégorie B
...
    MemberSummary:
      type: object
      required: [id, firstName, lastName, displayName, incomeCategory, status]
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        firstName:
          type: string
          minLength: 1
          maxLength: 100
          example: Amadou
        lastName:
          type: string
          minLength: 1
          maxLength: 100
          example: Diallo
        preferredName:
          type: string
          minLength: 1
          maxLength: 100
          example: Bah
        displayName:
          type: string
          minLength: 1
          maxLength: 201
          readOnly: true
          example: Amadou Diallo
        country:
          type: string
          minLength: 1
          maxLength: 100
          example: Guinée
        city:
          type: string
          minLength: 1
          maxLength: 100
          example: Conakry
        phone:
          type: string
          minLength: 7
          maxLength: 25
          pattern: '^\+?[0-9][0-9 ()-]{6,24}$'
          example: '+224 622 12 34 56'
        incomeCategory:
          $ref: '#/components/schemas/IncomeCategorySummary'
        associationFunction:
          type: string
          minLength: 1
          maxLength: 100
          description: Fonction descriptive sans effet automatique sur les droits applicatifs.
          example: Président
        status:
          $ref: '#/components/schemas/MemberStatus'
...
    Due:
      type: object
      required: [id, member, campaign, incomeCategorySnapshot, dueAmount, paidAmount, remainingAmount, status, paymentCount, currency]
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        member:
          $ref: '#/components/schemas/PersonSummary'
        campaign:
          $ref: '#/components/schemas/CampaignReference'
        incomeCategorySnapshot:
          $ref: '#/components/schemas/IncomeCategorySummary'
        dueAmount:
          type: integer
          format: int64
          minimum: 0
          readOnly: true
          example: 100000
        paidAmount:
          type: integer
          format: int64
          minimum: 0
          readOnly: true
          example: 50000
        remainingAmount:
          type: integer
          format: int64
          minimum: 0
          readOnly: true
          example: 50000
        status:
          $ref: '#/components/schemas/DueStatus'
        paymentCount:
          type: integer
          minimum: 0
          readOnly: true
          example: 1
        currency:
          $ref: '#/components/schemas/CurrencyCode'
...
    DuePage:
      type: object
      additionalProperties: false
      required: [items, page]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Due'
        page:
          $ref: '#/components/schemas/PageMetadata'
    CampaignReference:
      type: object
      additionalProperties: false
      required: [id, name, startDate, endDate, status]
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        name:
          type: string
          minLength: 1
          maxLength: 150
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        status:
          $ref: '#/components/schemas/CampaignStatus'
...
    SocialFundSummary:
      type: object
      required: [id, title, eventType, beneficiary, startDate, endDate, status, collectedAmount, contributorCount, contributionCount, currency]
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        title:
          type: string
          minLength: 1
          maxLength: 150
          example: Mariage de Fanta et Sékou
        eventType:
          $ref: '#/components/schemas/SocialEventType'
        beneficiary:
          type: string
          minLength: 1
          maxLength: 200
          example: Famille Camara
        startDate:
          type: string
          format: date
          example: '2026-09-05'
        endDate:
          type: string
          format: date
          example: '2026-09-28'
        status:
          $ref: '#/components/schemas/SocialFundStatus'
        targetAmount:
          type: integer
          format: int64
          minimum: 1
          description: Objectif facultatif en GNF.
          example: 7000000
        collectedAmount:
          type: integer
          format: int64
          minimum: 0
          readOnly: true
          example: 4750000
        remainingToTargetAmount:
          type: integer
          format: int64
          minimum: 0
          readOnly: true
          description: Présent uniquement lorsqu'un objectif est défini; vaut zéro si l'objectif est dépassé.
        progressRate:
          type: number
          format: double
          minimum: 0
          readOnly: true
          description: Présent uniquement lorsqu'un objectif est défini; peut dépasser 100.
          example: 67.9
        contributorCount:
          type: integer
          minimum: 0
          readOnly: true
          description: Nombre d'identités contributrices distinctes, membres par identifiant et externes par paire de prénom et nom normalisée dans cette cagnotte.
          example: 43
        contributionCount:
          type: integer
          minimum: 0
          readOnly: true
          example: 51
        currency:
          $ref: '#/components/schemas/CurrencyCode'
    SocialFund:
      unevaluatedProperties: false
      allOf:
        - $ref: '#/components/schemas/SocialFundSummary'
        - type: object
          properties:
            description:
              type: string
              maxLength: 1000
              example: Collecte de soutien à l'occasion du mariage.
...
    SocialFundReference:
      type: object
      additionalProperties: false
      required: [id, title, eventType, status]
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        title:
          type: string
          minLength: 1
          maxLength: 150
        eventType:
          $ref: '#/components/schemas/SocialEventType'
        status:
          $ref: '#/components/schemas/SocialFundStatus'
    ExternalContributor:
      type: object
      additionalProperties: false
      required: [firstName, lastName]
      properties:
        firstName:
          type: string
          minLength: 1
          maxLength: 100
          pattern: '.*\S.*'
          description: Prénom conservé comme instantané de la contribution externe.
        lastName:
          type: string
          minLength: 1
          maxLength: 100
          pattern: '.*\S.*'
          description: Nom conservé comme instantané de la contribution externe.
    Contribution:
      type: object
      additionalProperties: false
      required: [id, member, externalContributor, socialFund, amount, contributionDate, method, recordedBy, recordedAt, currency]
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        member:
          oneOf:
            - $ref: '#/components/schemas/PersonSummary'
            - type: 'null'
          description: Membre associé, nul pour une contribution externe.
        externalContributor:
          oneOf:
            - $ref: '#/components/schemas/ExternalContributor'
            - type: 'null'
          description: Identité externe associée, nulle pour une contribution membre.
        socialFund:
          $ref: '#/components/schemas/SocialFundReference'
        amount:
          type: integer
          format: int64
          minimum: 1
          example: 150000
        contributionDate:
          type: string
          format: date
          example: '2026-09-14'
        method:
          $ref: '#/components/schemas/PaymentMethod'
        recordedBy:
          $ref: '#/components/schemas/AuditActor'
        recordedAt:
          type: string
          format: date-time
          readOnly: true
          example: '2026-09-14T09:05:00Z'
        currency:
          $ref: '#/components/schemas/CurrencyCode'
    CreateContributionRequest:
      type: object
      additionalProperties: false
      required: [amount, contributionDate, method]
      description: La requête doit contenir exactement une identité, memberId ou externalContributor.
      properties:
        memberId:
          type: string
          format: uuid
          description: Identifiant du membre associé.
        externalContributor:
          $ref: '#/components/schemas/ExternalContributor'
        amount:
          type: integer
          format: int64
          minimum: 1
          description: Montant positif en GNF; aucune limite liée à l'objectif de la cagnotte.
        contributionDate:
          type: string
          format: date
        method:
          $ref: '#/components/schemas/PaymentMethod'
...
    ContributionCreationResponse:
      type: object
      additionalProperties: false
      required: [contribution, socialFund]
      properties:
        contribution:
          $ref: '#/components/schemas/Contribution'
        socialFund:
          $ref: '#/components/schemas/SocialFund'
    ContributionPage:
      type: object
      additionalProperties: false
      required: [items, page]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Contribution'
        page:
          $ref: '#/components/schemas/PageMetadata'
    PageMetadata:
      type: object
      additionalProperties: false
      required: [number, size, totalElements, totalPages]
      properties:
        number:
          type: integer
          minimum: 0
          readOnly: true
          example: 0
        size:
          type: integer
          minimum: 1
          maximum: 100
          readOnly: true
          example: 20
        totalElements:
          type: integer
          format: int64
          minimum: 0
          readOnly: true
          example: 91
        totalPages:
          type: integer
          minimum: 0
          readOnly: true
          example: 5
    ErrorCode:
      type: string
      description: Code stable exploitable par l'IHM sans analyser le message.
      enum:
        - VALIDATION_ERROR
        - DATE_RANGE_INVALID
        - AUTHENTICATION_REQUIRED
        - INVALID_CREDENTIALS
        - ACCESS_DENIED
        - RESOURCE_NOT_FOUND
        - BUSINESS_CONFLICT
        - DUPLICATE_CATEGORY_LABEL
        - MEMBER_ALREADY_INACTIVE
        - MEMBER_ALREADY_ACTIVE
        - INVALID_OPERATOR_CONFIGURATION
        - CAMPAIGN_NOT_EDITABLE
        - CAMPAIGN_NOT_READY
        - CAMPAIGN_START_DATE_NOT_REACHED
        - CAMPAIGN_ALREADY_OPEN
        - CAMPAIGN_CLOSED
        - CAMPAIGN_ALREADY_CLOSED
        - CAMPAIGN_NOT_OPEN
        - PAYMENT_EXCEEDS_REMAINING_AMOUNT
        - DUE_ALREADY_PAID
        - RESOURCE_CLOSED
        - SOCIAL_FUND_ALREADY_CLOSED
        - INTERNAL_ERROR
    FieldError:
      type: object
      additionalProperties: false
      required: [field, code, message]
      properties:
        field:
          type: string
          minLength: 1
          example: amount
        code:
          type: string
          minLength: 1
          example: MUST_BE_POSITIVE
        message:
          type: string
          minLength: 1
          example: Le montant doit être strictement positif.
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
          minLength: 1
          example: La requête ne peut pas être traitée.
        fieldErrors:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'